Add files via upload
This code collates data from google form and google sheet into python for further analysis and coding. A questionnaire that is created in google form using conditional marking and responses can be filled out by respondents. The data is collected and linked to a google sheet. The google sheet is further linked to python using the google sheet API which can be retrieved using Google service account. 

The data in the google sheet is coded into a dataframe. There are several conditions on the marking scheme of the questionnaire and the code is written as per that. Further, two additional dataframes are created subject to the conditions. One is the actual scorecard dataframe and the other one is an Improvement Suggestions Dataframe. 

The last part is about how the generated scorecard be mailed to respondents based on the mail ID entered by them in the google form and hence the google sheet.